Use UIMessage streams
