+++ src/client/index.ts
@@ -14,6 +14,7 @@ import type {
   StreamTextResult,
   ToolChoice,
   ToolSet,
+  UIMessage,
 } from "ai";
 import {
   generateObject,
@@ -72,10 +73,7 @@ import {
   syncStreams,
   type StreamingOptions,
 } from "./streaming.js";
-import {
-  mergeTransforms,
-  serializeTextStreamingPartsV5,
-} from "./textStreamParts.js";
+import { compressUIMessageChunks, mergeTransforms } from "./textStreamParts.js";
 import { createThread, getThreadMetadata } from "./threads.js";
 import type {
   ActionCtx,
@@ -605,7 +603,7 @@ export class Agent<
             {
               stream: opts.saveStreamDeltas,
               onAsyncAbort: call.fail,
-              compress: serializeTextStreamingPartsV5,
+              compress: compressUIMessageChunks<Tools>,
               abortSignal: args.abortSignal,
             },
             {
@@ -628,11 +626,6 @@ export class Agent<
         options?.saveStreamDeltas,
         streamTextArgs.experimental_transform,
       ),
-      onChunk: async (event) => {
-        await streamer?.addParts([event.chunk]);
-        // console.log("onChunk", chunk);
-        return streamTextArgs.onChunk?.(event);
-      },
       onError: async (error) => {
         console.error("onError", error);
         await call.fail(errorToString(error.error));
@@ -652,20 +645,21 @@ export class Agent<
         steps.push(step);
         const createPendingMessage = await willContinue(steps, args.stopWhen);
         await call.save({ step }, createPendingMessage);
-        if (!createPendingMessage) {
-          await streamer?.finish();
-        }
         return args.onStepFinish?.(step);
       },
     }) as StreamTextResult<
       TOOLS extends undefined ? AgentTools : TOOLS,
       PARTIAL_OUTPUT
     >;
+    const stream = streamer?.consumeStream(
+      result.toUIMessageStream<UIMessage<Tools>>(),
+    );
     if (
       (typeof options?.saveStreamDeltas === "object" &&
         !options.saveStreamDeltas.returnImmediately) ||
       options?.saveStreamDeltas === true
     ) {
+      await stream;
       await result.consumeStream();
     }
     const metadata: GenerationOutputMetadata = {
